requirements.txt
langchain==0.1.11
pypdf2==3.0.1
python-dotenv==1.0.1
streamlit==1.31.1
openai==1.13.3
faiss-cpu==1.8.0
altair==5.2.0
tiktoken==0.6.0

# For huggingface llms
huggingface-hub==0.21.3

# For instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2